update person-create and person-edit test
