+++ tests/e2e/cht-form/mock-config.js
@@ -4,7 +4,7 @@ const path = require('path');
 let server;
 const mockApp = express();
 mockApp.get('/', (req, res) => res.sendFile(path.join(__dirname, 'index.html')));
-mockApp.use(express.static(path.join(__dirname, '../../../webapp/dist/cht-form')));
+mockApp.use(express.static(path.join(__dirname, '../../../build/cht-form')));
 
 const startMockApp = (config, formName) => {
   mockApp.get('/form.js', (req, res) => res.sendFile(path.join(__dirname, `${config}/forms`, `${formName}.js`)));
